Aanpassingen 2 maart

+++ src/SimulationPackage/Action.java
@@ -1,14 +1,14 @@
 package SimulationPackage;
 
 public class Action implements Comparable<Action> {
-	
-	private final int time;
+
+	private final double time;
 	private final String toDo;
 	private final int chute;
 	private final int dest;
 	private final int shift;
-	
-	public Action (int time, String toDo, int chute, int dest, int shift) {
+
+	public Action (double time, String toDo, int chute, int dest, int shift) {
 		this.time = time;
 		this.toDo = toDo;
 		this.chute = chute;
@@ -16,7 +16,7 @@ public class Action implements Comparable<Action> {
 		this.shift = shift;
 	}
 
-	public int getTime() {
+	public double getTime() {
 		return time;
 	}
 
@@ -37,6 +37,7 @@ public class Action implements Comparable<Action> {
 	}
 
 
+
 	@Override
 	public String toString() {
 		return "Action [time=" + time + ", toDo=" + toDo + ", chute=" + chute + ", dest=" + dest + ", shift=" + shift
@@ -45,7 +46,7 @@ public class Action implements Comparable<Action> {
 
 	@Override
 	public int compareTo(Action action2) {
-		return (this.time - action2.time);
+		return (int) (this.time - action2.time);
 	}
 
 }
